requirements.txt
python-dotenv==0.21.0
spotipy==2.21.0
